src/containers/face-comparison/model/requirements.txt
requests==2.25.1
six==1.16.0
opencv-python-headless==4.5.3.56
numpy==1.19.5
Pillow==8.1.1
pyclipper==1.3.0
Shapely==1.7.1
urllib3==1.26.6
python-dateutil==2.8.2
certifi==2021.5.30
idna==2.10
chardet==4.0.0
onnxruntime==1.8.1
scikit-image